// packages/generated-files-bot/src/config.ts
// Copyright 2021 Google LLC
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     https://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export const CONFIGURATION_FILE_PATH = 'generated-files-bot.yml';

export interface ExternalManifest {
  type: 'json' | 'yaml';
  file: string;
  jsonpath: string;
}

export interface GeneratedFile {
  path: string;
  message?: string;
}

export interface Configuration {
  generatedFiles?: (string | GeneratedFile)[];
  externalManifests?: ExternalManifest[];
  ignoreAuthors?: string[];
}